[ArgPromotion] Remove checks for densely packed and padding accessed
The `canPaddingBeAccessed` function is removed because is not used
anymore. The `isDenselyPacked` static member should be moved into a
class that actually uses it.

Unfortunately, there is no test that actually uses these checks,
fp80.ll's @UseLongDoubleSafelyNoPromotion is protected from promotion
not by non-constant index in a GEP instruction but by pointer
invalidation.