[ArgPromotion] Involve the store instructions in the promotion
The `store` instructions are allowed for byval, however, these should
use the same checks as the `load` instructions do, i.e. be part of the
`ArgsToPromote` collection. For these instructions, the check for
interfering modifications can be disabled, though. The promotion
algorithm itself has been modified a lot: all the accesses (i.e. loads
and stores) are rewritten to an `alloca` (as it is currently done for
byval promotion, a single `alloca` instruction is emitted for a whole
argument). This allows the old code to work as it did before the
promotion. To optimize these new `alloca`s out, the `PromoteMemToReg`
function from `Transforms/Utils/PromoteMemoryToRegister.cpp` file is
invoked.