[ArgPromotion] Use an alloca for every argument part
In order to let the `PromoteMemToReg` promote as many `alloca`s as it
is possible, there should be no `GEP`s from the `alloca`s. To
eliminate the `GEP`s, its own `alloca` is generated for every
argument part and a single `alloca` for the whole argument is not used
anymore.

After this, all the end users of the original argument (the `load` and
`store` instructions) are retargeted to the emitted `alloca`s. The `GEP`
and `BitCast` instructions used before the promotion are eliminated from
the code using the dead code analysis that existed in the original
version of the ArgumentPromotion pass.