[ArgPromotion] Extract the `V` variable
In order to not use a lot of `U->getUser()` the variable `V` has been
extracted. Also, there is no place for returning `None` from the
`HandleEndUser` lambda for a `store` instruction, so the check of the
result of the `HandleEndUser` invocation for `store` instructions can be
handled in the exactly same way as for `load` ones.