[PGO] Instrument modules with at least a single function definition
It makes sense to instrument modules that contain at least a single
function definition only. When a module contains only function
declarations and, optionally, global variable definitions (a module for
the regular-LTO phase for thin-LTO when LTOUnit splitting is enabled,
for example), it makes no sense to instrument such module and moreover,
to introduce the `__llvm_profile_raw_version` variable into the module.

The patch also introduces the gmock-based unittest infrastructure for
PGO Instrumentation and adds some test cases to check whether the
instrumentation has taken place. The testing infrastructure for analysis
modules was borrowed from the `LoopPassManagerTest` unittest and
simplified a bit to handle module analysis passes only. Actually, we are
testing whether the result of a trivial analysis pass was invalidated by
the `PGOInstrumentGen` one: we exploit the fact the pass invalidates all
the analysis results after a module was instrumented.